Adapting fixes for the new update

+++ sources/sprite.h
@@ -329,8 +329,8 @@ namespace as1
         float previousPathY = 0.0f;
         float previousPathZ = 0.0f;
 
-        // Host-only high-FPS turn carry.  This lives outside the retail SPRITE
-        // layout and is used only when UNIT turning would quantize to zero.
+        // Fractional UNIT turn progress used only when a high frame rate makes
+        // the per-frame integer rotation step round to zero.
         float unitTurnSubstepCarry = 0.0f;
         int unitTurnSubstepSign = 0;
     };
@@ -1134,7 +1134,7 @@ namespace as1
         ANGLE m_direction;                       // +0x50
         std::uint32_t m_actionTimer = 0;         // +0x54
         int m_animationFrameTime = 0;            // +0x58
-        SpriteCommandStack m_commandStack;       // +0x5C..+0x6B on x86
+        SpriteCommandStack m_commandStack;       // +0x5C..+0x6B on 32-bit
         ActionAuxState* m_actionAuxState = nullptr; // +0x6C
         SPRITE* m_bestTargetSprite = nullptr;    // +0x70
 
@@ -1255,10 +1255,10 @@ namespace as1
         void DrawDebugOverlay() override;
 
         void drawRegionDebugBounds();
-        double regionScreenLeft() const noexcept;
-        double regionScreenTop() const noexcept;
-        double regionScreenRight() const noexcept;
-        double regionScreenBottom() const noexcept;
+        float regionScreenLeft() const noexcept;
+        float regionScreenTop() const noexcept;
+        float regionScreenRight() const noexcept;
+        float regionScreenBottom() const noexcept;
         void drawRegionTilesAndFog();
         int rebuildRegionFogRamp(int start, int end, int color);
